requirements.txt
discord.py
requests
pytz
PyGithub
psycopg2
asyncio
asyncpg
